+++ scrapy/spidermiddlewares/depth.py
