+++ scrapy/spidermiddlewares/httperror.py
