+++ scrapy/spidermiddlewares/offsite.py
