+++ tests/test_spidermiddleware_depth.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from scrapy.contrib.spidermiddleware.depth import DepthMiddleware
+from scrapy.spidermiddlewares.depth import DepthMiddleware
 from scrapy.http import Response, Request
 from scrapy.spider import Spider
 from scrapy.statscol import StatsCollector
@@ -37,7 +37,7 @@ class TestDepthMiddleware(TestCase):
 
         rdm = self.stats.get_value('request_depth_max', spider=self.spider)
         self.assertEquals(rdm, 1)
- 
+
     def tearDown(self):
         self.stats.close_spider(self.spider, '')
