+++ tests/test_spidermiddleware_offsite.py
@@ -4,7 +4,7 @@ from six.moves.urllib.parse import urlparse
 
 from scrapy.http import Response, Request
 from scrapy.spider import Spider
-from scrapy.contrib.spidermiddleware.offsite import OffsiteMiddleware
+from scrapy.spidermiddlewares.offsite import OffsiteMiddleware
 from scrapy.utils.test import get_crawler
 
 class TestOffsiteMiddleware(TestCase):
